Reapplying [DEV-10853] Committers using statecouchdb now store the CouchDB rev in statekeyindex
Also:
* Added safeguard in commit_handling.go to prevent panic if metadata not
found in the statekeyindex
* StateKeyIndex is now a sync.Locker

This reverts commit 1508312247de80a384cb61a27c30aca78c1c5285.

+++ core/ledger/kvledger/txmgmt/statedb/kvcache/kv_cache_provider.go
@@ -182,6 +182,8 @@ func (p *KVCacheProvider) GetLeveLDBIterator(namespace, startKey, endKey, ledger
 	if err != nil {
 		return nil, err
 	}
+	stateKeyIndex.Lock()
+	defer stateKeyIndex.Unlock()
 	return stateKeyIndex.GetIterator(namespace, startKey, endKey), nil
 
 }
@@ -233,28 +235,27 @@ func (p *KVCacheProvider) PrepareIndexUpdates(validatedTxOps []ValidatedTxOp, va
 }
 
 func (p *KVCacheProvider) ApplyIndexUpdates(indexUpdates []*statekeyindex.IndexUpdate, indexDeletes []statekeyindex.CompositeKey, ledgerID string) error {
-
-	//Add key index in leveldb
-	if len(indexUpdates) > 0 {
+	if len(indexUpdates) > 0 || len(indexDeletes) > 0 {
 		stateKeyIndex, err := statekeyindex.NewProvider().OpenStateKeyIndex(ledgerID)
 		if err != nil {
 			return err
 		}
-		err = stateKeyIndex.AddIndex(indexUpdates)
-		if err != nil {
-			return err
+		stateKeyIndex.Lock()
+		defer stateKeyIndex.Unlock()
+		//Add key index in leveldb
+		if len(indexUpdates) > 0 {
+			err = stateKeyIndex.AddIndex(indexUpdates)
+			if err != nil {
+				return err
+			}
 		}
-	}
 
-	// Delete key index in leveldb
-	if len(indexDeletes) > 0 {
-		stateKeyIndex, err := statekeyindex.NewProvider().OpenStateKeyIndex(ledgerID)
-		if err != nil {
-			return err
-		}
-		err = stateKeyIndex.DeleteIndex(indexDeletes)
-		if err != nil {
-			return err
+		// Delete key index in leveldb
+		if len(indexDeletes) > 0 {
+			err = stateKeyIndex.DeleteIndex(indexDeletes)
+			if err != nil {
+				return err
+			}
 		}
 	}
